Add config and promotion to ManageContract for embedded payment.
